requirements.txt
streamlit
transformers
torch
torchvision
pillow
scikit-learn
kagglehub
numpy
pandas